- Added 'Morph' action to the GuiContainer class, used in GuiDlgGameSelect. - Implemented focus tree and keyboard message distribution. - Lots of rewriting, clean-up etc. - Many other things...

+++ trunk/source/GuiDialogs/GuiDlgMenu.h
@@ -13,7 +13,7 @@ public:
     GuiDlgMenu(GuiContainer *parent, const char *name, int rows);
     virtual ~GuiDlgMenu();
 
-    void Initialize(const char **items, int num, int width);
+    void Initialize(const char **items, int num, float width);
     void CleanUp(void);
     SELRET DoModal(int *selected);
 private:
